refactor: restructure package hierarchy for Google integrations, rename classes/files for clarity
