+++ src/main/java/org/phong/zenflow/plugin/subdomain/nodes/builtin/integration/google/core/oauth.profile.schema.json
