+++ src/main/java/org/phong/zenflow/plugin/subdomain/nodes/builtin/integration/google/core/plugin.schema.json
