+++ src/main/java/org/phong/zenflow/plugin/subdomain/nodes/builtin/integration/google/drive/files/move/GoogleDriveMoveExecutor.java
@@ -7,7 +7,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.phong.zenflow.plugin.subdomain.execution.dto.ExecutionResult;
 import org.phong.zenflow.plugin.subdomain.execution.interfaces.PluginNodeExecutor;
 import org.phong.zenflow.plugin.subdomain.node.registry.PluginNode;
-import org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.google.drive.share.GoogleDriveServiceManager;
+import org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.google.drive.GoogleDriveServiceManager;
 import org.phong.zenflow.plugin.subdomain.resource.ScopedNodeResource;
 import org.phong.zenflow.workflow.subdomain.context.ExecutionContext;
 import org.phong.zenflow.workflow.subdomain.logging.core.NodeLogPublisher;
@@ -64,7 +64,7 @@ public class GoogleDriveMoveExecutor implements PluginNodeExecutor {
                 if (sourceFolderId == null || sourceFolderId.isBlank()) {
                     File current = drive.files().get(fileId).setFields("parents").execute();
                     if (current.getParents() != null && !current.getParents().isEmpty()) {
-                        sourceFolderId = current.getParents().stream().collect(Collectors.joining(","));
+                        sourceFolderId = String.join(",", current.getParents());
                     }
                 }
