+++ src/test/java/org/phong/zenflow/plugin/subdomain/registry/PluginTest.java
@@ -4,18 +4,18 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
 import org.phong.zenflow.plugin.infrastructure.persistence.entity.Plugin;
 import org.phong.zenflow.plugin.infrastructure.persistence.repository.PluginRepository;
-import org.phong.zenflow.plugin.subdomain.registry.definitions.CorePlugin;
-import org.phong.zenflow.plugin.subdomain.registry.definitions.IntegrationPlugin;
+import org.phong.zenflow.plugin.subdomain.nodes.builtin.core.CorePlugin;
+import org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.IntegrationPlugin;
 import org.phong.zenflow.plugin.subdomain.registry.definitions.TestPlugin;
-import org.phong.zenflow.plugin.subdomain.registry.definitions.GoogleDrivePlugin;
-import org.phong.zenflow.plugin.subdomain.registry.definitions.GoogleDocsPlugin;
+import org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.google.drive.GoogleDrivePlugin;
+import org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.google.docs.GoogleDocsPlugin;
 import org.phong.zenflow.plugin.subdomain.registry.profile.PluginProfileDescriptorRegistry;
 import org.phong.zenflow.plugin.subdomain.registry.profile.PluginProfileDescriptorDelegate;
 import org.phong.zenflow.plugin.subdomain.registry.settings.PluginSettingDescriptorDelegate;
 import org.phong.zenflow.plugin.subdomain.registry.settings.PluginSettingDescriptorRegistry;
 import org.phong.zenflow.plugin.subdomain.registry.settings.PluginSettingProvider;
-import org.phong.zenflow.plugin.subdomain.registry.definitions.google.GoogleOAuthAuthorizationService;
-import org.phong.zenflow.plugin.subdomain.registry.definitions.google.GoogleOAuthProfileDescriptor;
+import org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.google.core.GoogleOAuthAuthorizationService;
+import org.phong.zenflow.plugin.subdomain.nodes.builtin.integration.google.core.GoogleOAuthProfileDescriptor;
 import org.phong.zenflow.plugin.subdomain.registry.profile.PluginProfileDescriptor;
 import org.mockito.ArgumentCaptor;
 import org.springframework.beans.factory.annotation.Autowired;
